refactor: move Range to own file
