+++ riscv_analysis/src/parser/mod.rs
@@ -75,3 +75,6 @@ mod token_has_raw_text;
 
 mod position;
 pub use position::*;
+
+mod range;
+pub use range::*;
