+++ riscv_analysis/src/parser/token.rs
@@ -6,27 +6,7 @@ use uuid::Uuid;
 
 use crate::passes::DiagnosticLocation;
 
-use super::Position;
-
-#[derive(Debug, PartialEq, Clone, PartialOrd, Ord, Eq, Default, Serialize, Deserialize)]
-pub struct Range {
-    start: Position,
-    end: Position,
-}
-
-impl Range {
-    pub fn new(start: Position, end: Position) -> Range {
-        Range { start, end }
-    }
-
-    pub fn start(&self) -> &Position {
-        &self.start
-    }
-
-    pub fn end(&self) -> &Position {
-        &self.end
-    }
-}
+use super::Range;
 
 #[derive(Debug, PartialEq, Clone)]
 pub struct Token {
@@ -279,18 +259,6 @@ impl ToDisplayForTokenVec for Vec<Token> {
 }
 
 // implement display for Range
-impl std::fmt::Display for Range {
-    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        write!(
-            f,
-            "{}:{} - {}:{}",
-            self.start.one_idx_line(),
-            self.start.one_idx_column(),
-            self.end.one_idx_line(),
-            self.end.one_idx_column()
-        )
-    }
-}
 
 impl<T> DiagnosticLocation for With<T> {
     fn range(&self) -> Range {
